Moved Row struct into storage module
