+++ src/storage.rs
@@ -1,9 +1,21 @@
-use crate::sql::{Row, ROW_SIZE};
+use std::fmt;
+use std::io::{self, Read, Write};
+use std::mem::size_of;
 
 const MAX_PAGES: usize = 100;
 pub struct Table {
     num_rows: usize,
     pages: [Option<Box<Page>>; MAX_PAGES],
+}
+
+const USERNAME_LENGTH: usize = 32;
+const EMAIL_LENGTH: usize = 256;
+pub const ROW_SIZE: usize =
+    size_of::<u32>() + size_of::<[u8; USERNAME_LENGTH]>() + size_of::<[u8; EMAIL_LENGTH]>();
+pub struct Row {
+    id: u32,
+    username: [u8; USERNAME_LENGTH],
+    email: [u8; EMAIL_LENGTH],
 }
 
 const PAGE_SIZE: usize = 4096;
@@ -80,5 +92,83 @@ impl Page {
 impl Default for Page {
     fn default() -> Self {
         Page::new()
+    }
+}
+
+impl Row {
+    pub fn write_to_buffer(&self, dst: &mut dyn Write) -> io::Result<()> {
+        let id_bytes = self.id.to_ne_bytes();
+        dst.write(id_bytes.as_slice())?;
+        dst.write(self.username.as_slice())?;
+        dst.write(self.email.as_slice())?;
+        Ok(())
+    }
+
+    pub fn from_buffer(src: &mut dyn Read) -> io::Result<Row> {
+        let mut id_bytes = [0u8; size_of::<u32>()];
+        src.read_exact(&mut id_bytes)?;
+        let id = u32::from_ne_bytes(id_bytes);
+
+        let mut username = [0u8; USERNAME_LENGTH];
+        src.read_exact(&mut username)?;
+
+        let mut email = [0u8; EMAIL_LENGTH];
+        src.read_exact(&mut email)?;
+
+        Ok(Row {
+            id,
+            username,
+            email,
+        })
+    }
+
+    pub fn new(id: u32, username: &str, email: &str) -> Row {
+        let mut username_buffer = [0u8; USERNAME_LENGTH];
+        let mut email_buffer = [0u8; EMAIL_LENGTH];
+
+        username_buffer
+            .as_mut_slice()
+            .write(username.as_bytes())
+            .unwrap();
+        email_buffer.as_mut_slice().write(email.as_bytes()).unwrap();
+
+        Row {
+            id,
+            username: username_buffer,
+            email: email_buffer,
+        }
+    }
+}
+
+impl fmt::Display for Row {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "{} {} {}",
+            self.id,
+            String::from_utf8_lossy(&self.username),
+            String::from_utf8_lossy(&self.email)
+        )
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::Row;
+
+    #[test]
+    fn serialization_and_deserialization_are_reverses_of_eachother() {
+        let mut test_memory = [0u8; 512];
+
+        let row = Row::new(3, "Test User", "test@tester.com");
+
+        row.write_to_buffer(&mut test_memory.as_mut_slice())
+            .unwrap();
+
+        let deserialized_row = Row::from_buffer(&mut test_memory.as_slice()).unwrap();
+
+        assert_eq!(row.id, deserialized_row.id);
+        assert_eq!(row.username, deserialized_row.username);
+        assert_eq!(row.email, deserialized_row.email);
     }
 }
